+++ pages/_campus/edit.vue
@@ -224,6 +224,7 @@ import ecGpsPoint from '~/components/form/gps-point.vue';
 import searchCategories from '~/components/form/search-categories.vue';
 import weekdaysSelect from '~/components/form/weekdays.vue';
 import rideDuration from '~/components/form/ride-duration.vue';
+import formatCoordinates from '~/helpers/format-coordinates';
 
 const EDITABLE_FIELDS = 'id,name,location,phone(drivers,everybody),categories(id,label),'
     + 'information,timezone,workedDays,workedHours,defaultRideDuration,defaultReservationScope';
@@ -284,10 +285,17 @@ export default {
   methods: {
     async edit(campus) {
       let data = {};
+      const formattedCampus = {
+        ...campus,
+        location: {
+          coordinates: formatCoordinates(campus.location.coordinates),
+        },
+      };
+
       if (this.campus.id) {
-        ({ data } = (await this.$api.campuses.patchCampus(campus.id, campus, EDITABLE_FIELDS)));
+        ({ data } = (await this.$api.campuses.patchCampus(campus.id, formattedCampus, EDITABLE_FIELDS)));
       } else {
-        ({ data } = (await this.$api.campuses.postCampus(campus, EDITABLE_FIELDS)));
+        ({ data } = (await this.$api.campuses.postCampus(formattedCampus, EDITABLE_FIELDS)));
       }
 
       this.$store.commit('context/setCampus', data);